Cleanup some IntelliSense warnings.

+++ drone_move.c
@@ -797,7 +797,7 @@ qboolean CheckYawStep (edict_t *self, float minyaw, float maxyaw, float dist)
 	AngleCheck(&maxyaw);
 
 	// calculate the maximum yaw variance
-	max = 360 - fabs(minyaw - maxyaw);
+	max = 360 - fabsf(minyaw - maxyaw);
 
 	// we will start at the minimum yaw angle and move towards maxyaw
 	yaw = minyaw;
@@ -805,7 +805,7 @@ qboolean CheckYawStep (edict_t *self, float minyaw, float maxyaw, float dist)
 	for (i = 0; i < max; i += 30) 
 	{
 		// if we changed course a while ago, then try a partial step 50% of the time
-		if (level.time - self->monsterinfo.bump_delay > 1.0 || random() < 0.5)
+		if (level.time - self->monsterinfo.bump_delay > 1.0f || random() < 0.5f)
 		{
 			yaw += i;
 			AngleCheck(&yaw);
@@ -1131,18 +1131,18 @@ qboolean M_walkmove (edict_t *ent, float yaw, float dist)
 
 			// delta between monster yaw and wall yaw should be no more than 90 degrees
 			// else, turn wall angles around 180 degrees
-			if (fabs(cl_yaw-angles[YAW]) > 90)
+			if (fabsf(cl_yaw-angles[YAW]) > 90)
 				angles[YAW]+=180;
 			ValidateAngles(angles);
 
 			// possible escape angle 1
 			angle1 = angles[YAW]+90;
 			AngleCheck(&angle1);
-			delta1 = fabs(angle1-cl_yaw);
+			delta1 = fabsf(angle1-cl_yaw);
 			// possible escape angle 2
 			angle2 = angles[YAW]-90;
 			AngleCheck(&angle2);
-			delta2 = fabs(angle2-cl_yaw);
+			delta2 = fabsf(angle2-cl_yaw);
 
 			// take the shorter route
 			if (delta1 > delta2)
